Refactor and fix skillService
