+++ app/src/main/java/edu/kmaooad/capstone23/competences/events/SkillDeleted.java
@@ -2,12 +2,10 @@ package edu.kmaooad.capstone23.competences.events;
 
 import edu.kmaooad.capstone23.competences.dal.Skill;
 
-import java.util.Optional;
-
 public class SkillDeleted {
     private Skill skill;
 
-    public SkillDeleted(Optional<Skill> skill) {
+    public SkillDeleted(Skill skill) {
         this.skill = skill;
     }
