+++ app/src/main/java/edu/kmaooad/capstone23/competences/services/impl/SkillServiceImpl.java
@@ -3,18 +3,49 @@ package edu.kmaooad.capstone23.competences.services.impl;
 import edu.kmaooad.capstone23.competences.dal.Skill;
 import edu.kmaooad.capstone23.competences.dal.SkillsRepository;
 import edu.kmaooad.capstone23.competences.services.SkillService;
-import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import java.util.Optional;
 import org.bson.types.ObjectId;
+import java.util.List;
+import java.util.Optional;
 
-@ApplicationScoped
-public class SkillServiceImpl implements SkillService {
-  @Inject
-  private SkillsRepository repo;
+public class SkillServiceImpl implements SkillService{
+    @Inject
+    private SkillsRepository skillsRepository;
 
-  @Override
-  public Optional<Skill> findByIdOptional(ObjectId id) {
-    return repo.findByIdOptional(id);
-  }
+    @Override
+    public Skill findById(String id) {
+        return skillsRepository.findById(id);
+    }
+
+    @Override
+    public Skill insert(Skill skill) throws IllegalArgumentException {
+        if(skill.parentSkill != null && skillsRepository.findByIdOptional(skill.parentSkill.toString()).isEmpty())
+            throw new IllegalArgumentException("Parent has unknown id");
+        return skillsRepository.insert(skill);
+    }
+
+    @Override
+    public Optional<Skill> findByIdOptional(String id) {
+        return skillsRepository.findByIdOptional(id);
+    }
+
+    @Override
+    public void delete(Skill skill) {
+        skillsRepository.deleteSkill(skill);
+    }
+
+    @Override
+    public List<Skill> findChildRepositories(ObjectId parentSkill) {
+        return skillsRepository.findChildRepositories(parentSkill);
+    }
+
+    @Override
+    public Skill update(Skill skill) throws  IllegalArgumentException {
+        if(skill.parentSkill != null && skillsRepository.findByIdOptional(skill.parentSkill.toString()).isEmpty())
+            throw new IllegalArgumentException("Parent has unknown id");
+        if(skill.id.equals(skill.parentSkill))
+            throw new IllegalArgumentException("Parent id and id are equal");
+        skillsRepository.modify(skill);
+        return skill;
+    }
 }
